debugging good

+++ inc/malloc.h
@@ -49,7 +49,7 @@ struct	s_pages_control
 	char			errors;
 
 	// // size_t count;
-	int debug;
+	size_t debug;
 };
 
 struct s_memory_ctrl
